# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IPRP Backend API",
    "version": "2.0.0",
    "description": "Server-enforced firm tenancy. BE-14 billing is deferred and has no endpoint."
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Replace only with an approved environment."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/invitations/{token}": {
      "get": {
        "operationId": "invitationDetails",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/token"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Invitation"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/provisioning/firm": {
      "post": {
        "operationId": "provisionFirm",
        "x-roles": [
          "verified identity"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Firm"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Firm"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/admin/invitations": {
      "get": {
        "operationId": "listAdminInvitations",
        "x-roles": [
          "admin"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Invitation"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "issueInvitation",
        "x-roles": [
          "admin"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/InvitationIssue"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Invitation"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/users/{id}/role": {
      "patch": {
        "operationId": "changeUserRole",
        "x-roles": [
          "admin"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Role"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Me"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/auth/invitations/{token}/redeem": {
      "post": {
        "operationId": "redeemInvitation",
        "parameters": [
          {
            "$ref": "#/components/parameters/token"
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/components/responses/Firm"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "410": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/provisioning/organization-intents": {
      "post": {
        "operationId": "startOrganizationIntent",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/OrganizationIntent"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Firm"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "listAdminUsers",
        "x-roles": [
          "admin"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Invitation"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/admin/invitations/{invitationId}/resend": {
      "post": {
        "operationId": "resendInvitation",
        "x-roles": [
          "admin"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Invitation"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/admin/invitations/{invitationId}": {
      "delete": {
        "operationId": "revokeInvitation",
        "x-roles": [
          "admin"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked."
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/admin/users/{userId}/role": {
      "patch": {
        "operationId": "changeAdminUserRole",
        "x-roles": [
          "admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Role"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Invitation"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Verified Supabase access token; refresh/session ownership remains with Supabase."
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "token": {
        "name": "token",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
...
    },
    "requestBodies": {
      "Firm": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "intentToken"
              ],
              "properties": {
                "intentToken": {
                  "type": "string",
                  "minLength": 40,
                  "maxLength": 128
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
      "InvitationIssue": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "fullName",
                "email",
                "role"
              ],
              "properties": {
                "fullName": {
                  "type": "string",
                  "maxLength": 200
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "role": {
                  "enum": [
                    "admin",
                    "attorney",
                    "viewer"
                  ]
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
      "Role": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "role"
              ],
              "properties": {
                "role": {
                  "enum": [
                    "admin",
                    "attorney",
                    "viewer"
                  ]
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
      "OrganizationIntent": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email",
                "firmName"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "firmName": {
                  "type": "string",
                  "maxLength": 200
                }
              },
              "additionalProperties": false
            }
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Stable safe error; never contains credentials, stacks, storage keys, or raw upstream bodies.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Me": {
        "description": "Resolved membership.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Me"
            }
          }
        }
      },
      "Firm": {
        "description": "Firm/user provisioning result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Firm"
            }
          }
        }
      },
      "Invitation": {
        "description": "Safe invitation result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Invitation"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Me": {
        "type": "object",
        "required": [
          "userId",
          "email",
          "role",
          "firmId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "enum": [
              "admin",
              "attorney",
              "viewer"
            ],
            "nullable": true
          },
          "firmId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "Firm": {
        "type": "object",
        "additionalProperties": true
      },
      "Invitation": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}